Check the returned status code for more explicit error handling
By using the `Result<(), Box<dyn std::error::Error>>` we can return any
type of error from the function. It will be a "boxed" error.

`send?` will bubble up any errors coming from `reqwest` calls.

If the response is OK, we return `Ok(())`.

If the response is not OK, we return an `Err` with the status code
converted into an error using `into()`.

This means we can have more fine grained control over the errors we
return in the future.